pyproject.toml
[project]
name = "slack-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "load-dotenv>=0.1.0",
    "openai>=2.17.0",
    "slack-bolt>=1.27.0",
]
